refactor: zio-http-logging package name
