+++ zio-http/src/main/scala/zio/http/netty/server/ServerChannelInitializer.scala
@@ -17,7 +17,7 @@ import zio.http.netty.Names
 import zio.http.netty.server.ServerChannelInitializer.log
 import zio.http.service.Log
 import zio.http.service.logging.LogLevelTransform._
-import zio.logging.LogLevel
+import zio.http.logging.LogLevel
 import zio.stacktracer.TracingImplicits.disableAutoTrace // scalafix:ok;
 
 /**
